Print out required connection status, removed zergs and health status of zergs also removed commited out lines and print statements

+++ zergmapper.c
@@ -7,26 +7,25 @@
 #include "splay.h"
 #include "lgraph.h"
 #include "haversine.h"
-//#include "min_heap.h"
 #include "zerg_connect.h"
 
 #define DIRECTED 1
 #define UNDIRECTED 0
-#define MAX 99999.99
 
 void display_zerg(const void * data);
 void display_removed(const void * data);
 void get_health(struct node * root, int health);
+
 void initialize_graph(graph_ptr graph, struct node * zerg, struct tree * pcaps);
 void get_all_edges(graph_ptr graph, struct node * zerg, struct node * root);
-void adding_nodes(graph_ptr graph, struct node * zerg, struct node * root);
 double get_distance(double src_lat, double src_long, double src_alt, double dest_lat, double dest_long, double dest_alt);
-void find_reachable(graph_ptr graph, const int start, bool reach[]);
-
-void shortest_path( graph_ptr graph, int start_vertex, int graph_nodes );
-void shortest_path_v2(graph_ptr graph, int start_node, int graph_nodes);
-void zerg_connected(graph_ptr graph, int start, int end, int num_nodes);
+void adding_nodes(graph_ptr graph, struct node * zerg, struct node * root);
 struct node * removeNode(graph_ptr graph, int node, struct node * root, int * count );
+
+void find_reachable(graph_ptr graph, const int start, bool reach[]);
+void zerg_connected(graph_ptr graph, int start, int end, int num_nodes);
+int check_first_vertex(graph_ptr graph, int nodes);
+int check_last_vertex(graph_ptr graph, int nodes);
 
 int main(int argc, char **argv)
 {
@@ -42,33 +41,33 @@ int main(int argc, char **argv)
 			break;
 	}
 	
-	printf("min_health: %d \n", min_health);
-	
 	struct tree * pcap_nodes = create_tree();
 	pcap_nodes->head = decode(argc, argv, &pcap_nodes->count);
 	int nodes = pcap_nodes->count;
 	
-	//printf("number of nodes  = %d \n", pcap_nodes->count);
-	if ( pcap_nodes->head != NULL )
-	{
-		printf("\nsuccesful call to decode\n");
-	}
-	else
-	{
-		printf("fail to call decode\n");
-	}
+	printf("Display zergs health percentage from data received from input below %d %%\n", min_health);
+	get_health(pcap_nodes->head, min_health);  
 	
-	//printf("\n\nprinting nodes\n\n");
-	preOrder(pcap_nodes->head, display_zerg);
-	get_health(pcap_nodes->head, min_health);  //-> HOw to handle zergs with 0 hit points.
+	//Create and initialize graph with data from tree data structure.
 	graph_ptr dir_graph = createGraph(nodes, DIRECTED);
 	initialize_graph(dir_graph, pcap_nodes->head, pcap_nodes);
 	
-	remove_tree(pcap_nodes);
-	printf("\nDIRECTED GRAPH\n");
-	displayGraph(dir_graph);
-	
-	
+	// Determine if there is enough nodes to build graph
+	if ( nodes <= 2 )
+	{
+		printf("not enough nodes to build a graph from pcaps\n");
+		preOrder(pcap_nodes->head, display_zerg);
+		remove_tree(pcap_nodes);
+		return 2;
+	}
+
+	int last_vertex;
+	int first_vertex;
+	// Get the first index position.
+	first_vertex = check_first_vertex(dir_graph, nodes);
+	//Find the last vertex in the graph
+	last_vertex = check_last_vertex(dir_graph, nodes);
+	//printf("first = %d\nlast = %d\n", first_vertex, last_vertex);
 	
 	/* GETTING THE REACHABILITY of ALL NODES */
 	int i;
@@ -83,36 +82,35 @@ int main(int argc, char **argv)
 	//Create a tree to hold the removed zergs src_id;
 	struct tree * dead_zerg = create_tree();
 	dead_zerg->head = NULL;
-	//printf("getting reachability of all nodes from a single point\n");
-	find_reachable(dir_graph, 0, reach);
-	printf("after\n");
+	
+	// Remove zergs that are not reachable from a single source.
+	find_reachable(dir_graph, first_vertex, reach);
 	for (i = 0; i < nodes; i++)
 	{
-		//printf("reach[%d]=%d \n", i, reach[i]);
 		if ( reach[i] == 0)
 		{
-			//printf("node = unreachable\n");
-			//If node is unreachable, remove it from the graph -> add to BST.
-			dead_zerg->head = removeNode(dir_graph, i, dead_zerg->head, &dead_zerg->count);
+			if ( !(dir_graph->adjListArr[i].head == NULL) )
+			{
+				//If node is unreachable, remove it from the graph -> add to BST.
+				dead_zerg->head = removeNode(dir_graph, i, dead_zerg->head, &dead_zerg->count);
+			}
 		}
-	} 
-	printf("\nUPDATED after Reachable check GRAPH");
-	displayGraph(dir_graph);
-	printf("\n");
+	}
+	// Release the malloc'd space for reach array
+	free(reach);
 	
+	// Remove zergs with only a single edge
 	for ( i = 0; i < nodes; i++)
 	{
-		//printf("Node [%d]: %d edges \n", i, dir_graph->adjListArr[i].num_members);
 		if(  dir_graph->adjListArr[i].num_members < 2)
 		{
 			if ( !(dir_graph->adjListArr[i].head == NULL) )
+			{
 				dead_zerg->head = removeNode(dir_graph, i, dead_zerg->head, &dead_zerg->count);
+			}
 		}
 	}
-	
-	printf("Number of nodes in removed_nodes: %d \n", dead_zerg->count );
-	printf("Percentage: %.2f \n", (dead_zerg->count / (double)nodes) );
-	
+	// Check if the number of zergs removed is over 50% of the total number of zergs.
 	if (  ( (dead_zerg->count / (double)nodes ) * 100) > 50.0)
 	{
 		printf("Removed more than 50%% of zergs : Not Zerg Connected \n");
@@ -120,39 +118,77 @@ int main(int argc, char **argv)
 		return 1;
 	}
 	
-	printf("\nUPDATED before path finding GRAPH");
-	displayGraph(dir_graph);
-	printf("\n");
-	
-	int last_vertex;
-	//Find the last vertex in the graph
+	int updated_node_count = 0;
+	// GEt an updated number of valid nodes in the graph prior to path-finding.
 	for ( int x = 0; x < nodes; x++)
 	{
-		if ( !(dir_graph->adjListArr[x].head == NULL) )
+		if ( (dir_graph->adjListArr[x].head != NULL) )
 		{
-			last_vertex = x;
+			updated_node_count++;
 		}
 	}
 	
-	//Attempt to determine zerg connected by getting 2 disjoint paths from index 0 to last useable vertex;
-	zerg_connected(dir_graph, 0, last_vertex, nodes);
+	// Check the first vertex in the graph
+	first_vertex = check_first_vertex(dir_graph, nodes);
+	// Check the last vertex in the graph
+	last_vertex = check_last_vertex(dir_graph, nodes);
 	
-	
-	free(reach);
-	destroyGraph(dir_graph);
-	
+	if ( updated_node_count == 0 )
+	{
+		printf("No zergs in the graph \n");
+	}
+	else
+	{
+		//Attempt to determine zerg connected by getting 2 disjoint paths from index 0 to last useable vertex;
+		zerg_connected(dir_graph, first_vertex, last_vertex, updated_node_count);
+	}
+
+	//Output zergs that were removed from the graph
 	preOrder(dead_zerg->head, display_removed);
-	remove_tree(dead_zerg);	
+	
+	//Clean up data structures and release memory
+	destroyGraph(dir_graph);
+	remove_tree(pcap_nodes);
+	remove_tree(dead_zerg);
+	if ( dead_zerg == NULL)
+		free(dead_zerg);
+	
 	return 0;
 }
 
+int check_first_vertex(graph_ptr graph, int nodes)
+{
+	int first_vertex;
+	for ( int x = 0; x < nodes; x++)
+	{
+		if ( (graph->adjListArr[x].head != NULL) )
+		{
+			first_vertex = x;
+			break;
+		}
+	}
+	return first_vertex;
+}
+
+int check_last_vertex(graph_ptr graph, int nodes)
+{
+	int last_vertex;
+	
+	for ( int x = 0; x < nodes; x++)
+		{
+			if ( !(graph->adjListArr[x].head == NULL) )
+			{
+				last_vertex = x;
+			}
+		}
+		
+	return last_vertex;
+}
 void get_health(struct node * root, int min_health)
 {
-	//printf("here in get_health \n");
 	if ( root != NULL)
 	{
-		double health_pct = (double)((struct zerg*)root->key)->health.hit_points / (double)((struct zerg*)root->key)->health.max_points;
-		//printf("health_pct = %f \n", health_pct);
+		double health_pct = ((struct zerg*)root->key)->health.hit_points / (double)((struct zerg*)root->key)->health.max_points;
 		if ( ( health_pct * 100) <= min_health )
 			//Output the health as a percentage.
 			printf("zerg: %3d has health of %0.2f%% \n", ((struct zerg*)root->key)-> srcID, health_pct * 100);
@@ -166,17 +202,10 @@ void get_health(struct node * root, int min_health)
 void find_reachable(graph_ptr graph, const int current, bool reach[])
 {
 	reach[current] = true;
-	//printf("current = %d \n", current);
-	//printf("
 	adjlist_node_ptr start = graph->adjListArr[current].head;
-	//printf("here  %d \n", start->vertex);
-	//printf("here  %u \n", start->src_ID);
-	//printf("here  %f \n", start->distance);
 	
 	for (;start;start = start->next) 
 	{
-		//printf("start->vertex = %d \n", start->vertex);
-		
 		if( !reach[start->vertex] ) 
 		{
 			find_reachable(graph,start->vertex,reach);
@@ -187,9 +216,6 @@ void find_reachable(graph_ptr graph, const int current, bool reach[])
 //Function to get haversine equation results.
 double get_distance(double src_lat, double src_long, double src_alt, double dest_lat, double dest_long, double dest_alt)
 {
-	//printf("\n\nin get_distance\n");
-	//printf("zerg_src->lat %f \n", src_lat );
-	//printf("zerg_dest->lat %f \n", dest_lat );
 	return haversine_dist(src_lat, src_long, src_alt, dest_lat, dest_long, dest_alt);
 }
 
@@ -198,8 +224,6 @@ void initialize_graph(graph_ptr graph, struct node * zerg, struct tree * pcaps)
 {
 	if ( zerg != NULL)
 	{
-		//display_zerg(zerg->key);
-		//printf("zerg_id = %u \n", ((struct zerg*)zerg->key)->srcID );
 		get_all_edges(graph, zerg, pcaps->head);
 		initialize_graph(graph, zerg->left, pcaps);
 		initialize_graph(graph, zerg->right, pcaps);
@@ -209,7 +233,6 @@ void initialize_graph(graph_ptr graph, struct node * zerg, struct tree * pcaps)
 
 void get_all_edges(graph_ptr graph, struct node * zerg, struct node * root)
 {
-	//printf("here in getting all edges \n");
 	if ( root != NULL)
 	{
 		adding_nodes(graph, zerg, root);
@@ -223,23 +246,27 @@ void get_all_edges(graph_ptr graph, struct node * zerg, struct node * root)
 void adding_nodes(graph_ptr graph, struct node * zerg, struct node * root)
 {
 	
-	//printf("\n\nsrc\n");
-	//printf("zerg_src->lat %f \n", ((struct zerg*)zerg->key)->position.latitude.value );
-	//printf("zerg_src->long %f \n", ((struct zerg*)zerg->key)->position.longitude.value );
-	
 	double src_lat = ((struct zerg*)zerg->key)->position.latitude.value ;
 	double src_long = ((struct zerg*)zerg->key)->position.longitude.value;
 	double src_alt = ((struct zerg*)zerg->key)->position.altitude.value;
 	int src_vertex = ((struct zerg*)zerg->key)->number;
 	
-	//printf("\n\ndest\n");
-	//printf("zerg_dest->lat %f \n", ((struct zerg*)root->key)->position.latitude.value );
-	//printf("zerg_dest->long %f \n", ((struct zerg*)root->key)->position.longitude.value );
+	// Do not process data with gps data out of range
+	if ( abs(src_lat) > 90 || abs(src_long) > 180)
+	{
+		return;
+	}
 	
 	double dest_lat = ((struct zerg*)root->key)->position.latitude.value;
 	double dest_long = ((struct zerg*)root->key)->position.longitude.value;
 	double dest_alt = ((struct zerg*)root->key)->position.altitude.value;
 	int dest_vertex = ((struct zerg*)root->key)->number;
+	
+	// Do not process data with gps data out of range
+	if ( abs(dest_lat) > 90 || abs(dest_long) > 180)
+	{
+		return;
+	}
 	
 	double distance = 0.00;
 	if ( root != NULL)
@@ -249,14 +276,11 @@ void adding_nodes(graph_ptr graph, struct node * zerg, struct node * root)
 		//If distance meets minimum distance ( 1.25 yds * 0.9144) converts to meters.
 		if ( distance  > ( 1.25  * 0.9144)) 
 		{
-			//printf("distance = %f \n", distance);
 			//If distance does not exceed max distance of 15 meters.
 			if (distance > 15)
 			{
-				//printf("over 15 mdistance = %f \n", distance);
 				return;
 			}
-			//printf("distance is withing range!!!! \n");
 			addEdge(graph, src_vertex,((struct zerg*)zerg->key)->srcID, dest_vertex,((struct zerg*)root->key)->srcID, distance);
 		}
 	}
@@ -265,15 +289,13 @@ void adding_nodes(graph_ptr graph, struct node * zerg, struct node * root)
 
 struct node * removeNode(graph_ptr graph, int node, struct node * root, int * count )
 {
-	//printf("inserting node %d to be removed\n", node);
 	//Create a node for the node to be inserted
-	root = insert(root, node, sizeof(int), get_root_srcID);
+	root = insert(root, graph->adjListArr[node].src_ID, sizeof(int), get_root_srcID);
 	(*count)++;
 	//Assign the data to the node
-	*((uint16_t*)(root->key)) = node;
+	*((uint16_t*)(root->key)) = graph->adjListArr[node].src_ID;
 	while ( graph->adjListArr[node].head != NULL)
 	{
-		//printf("edge vertex to del: [%d] \n",  graph->adjListArr[node].head->vertex);
 		int edge_node = graph->adjListArr[node].head->vertex;
 		removeEdge(graph, node, edge_node);
 		removeEdge(graph, edge_node, node);
@@ -287,7 +309,7 @@ void display_zerg(const void * data)
 	printf("zerg_id: %u \n", ((struct zerg*)data)->srcID );
 	printf("zerg_lat: %06f \n",  ((struct zerg*)data)->position.latitude.value );
 	printf("zerg_long: %06f \n",  ((struct zerg*)data)->position.longitude.value );
-	printf("zerg_alt: %06f \n",  ((struct zerg*)data)->position.altitude.value );
+	printf("zerg_alt: %06f \n",  ((struct zerg*)data)->position.altitude.value  * 1.8288);
 	printf("\n");
 }
 void display_removed(const void * data)
